api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Guillaume Desforges <guillaume.desforges.pro@gmail.com>"]
readme = "README.md"
include = [ "./src", "./tests" ]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.2"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
requests = "^2.31.0"
python-dotenv = "^1.0.0"
indexer = "*"
langchain = "^0.0.184"
sentence-transformers = "^2.2.2"
pymilvus = "^2.2.8"
atlassian-python-api = "^3.38.0"
lxml = "^4.9.2"
rwkv = "^0.7.4"
tokenizers = "^0.13.3"
ruff = "^0.0.270"


[tool.poetry.group.dev.dependencies]
indexer = {path = "../indexer", develop = true}
black = "^23.3.0"
isort = "^5.12.0"
flake8 = "^6.0.0"
pyright = "^1.1.309"
httpx = "^0.24.1"
pytest = "^7.3.1"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F", "I", "N", "A", "ISC", "PLR", "SIM", "RET", "Q", "ICN", "ARG", "ERA", "C", "ANN", "B", "C4", "G", "PIE", "PYI", "PT", "PD", "PL", "TRY", "NPY", "RUF"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "G", "I", "N", "Q", "S", "T", "W", "ANN", "ARG", "BLE", "COM", "DJ", "DTZ", "EM", "ERA", "EXE", "FBT", "ICN", "INP", "ISC", "NPY", "PD", "PGH", "PIE", "PL", "PT", "PTH", "PYI", "RET", "RSE", "RUF", "SIM", "SLF", "TCH", "TID", "TRY", "UP", "YTT"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py310"


[tool.ruff.mccabe]
# Unlike Flake8, default to a complexity level of 10.
max-complexity = 10